+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 11/7/25
 
 - added `ui_entity.h`
+- `dl_material_activate` now binds the texture
+- `dl_entity_update` no longer binds texture, which is now done by `dl_material_activate`
+- added an order to how `mat4_custom` multiplies position, rotation, and scale with `Mat4_TransformationOrder`
 
 ## 10/7/25